authservice/requirements-dev.txt
asgiref==3.5.0
cffi==1.15.1
cryptography==37.0.2
Django==4.0.3
django-cors-headers==3.11.0
python-dotenv==0.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
ez-setup==0.9
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5